refactor(Problem1): 클래스 분리
