+++ src/main/java/onboarding/problem1/Constant.java
@@ -1,0 +1,5 @@
+package onboarding.problem1;
+
+public class Constant {
+    public static final int EXCEPTION = -1; // 시스템 상수
+}
